backend/requirements.txt
anyio==3.3.4
argon2-cffi==21.1.0
asgiref==3.4.1
asyncpg==0.24.0
cffi==1.15.0
click==8.0.3
databases==0.5.3
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
orm==0.2.1
pycparser==2.20
pydantic==1.8.2
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.26
starlette==0.16.0
typesystem==0.4.0
typing-extensions==3.10.0.2
uvicorn==0.15.0
